[16] - Modificar test para usar test data builder
