--- notificationsManager/Operators/Repositories/OperatorBuilder.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace NotificationsManager\Operators\Repositories;
-
-class OperatorBuilder
-{
-}
