+++ notificationsManager/Operators/Tests/Unit/ApiOperatorsDataSourceTest.php
@@ -6,8 +6,8 @@ use Exception;
 use GuzzleHttp\Client;
 use GuzzleHttp\Psr7\Response;
 use Mockery\Container;
-use NotificationsManager\ApiOperatorsDataSource;
-use NotificationsManager\Operators\Database\Entities\Operator;
+use NotificationsManager\Operators\ApiOperatorsDataSource;
+use NotificationsManager\Operators\Tests\OperatorTestDataBuilder;
 use PHPUnit\Framework\Attributes\Test;
 use Tests\TestCase;
 
@@ -15,47 +15,6 @@ class ApiOperatorsDataSourceTest extends TestCase
 {
     private Client $client;
     private ApiOperatorsDataSource $apiDataSource;
-    private $operatorDataApi = '[
-        {
-            "sequenceNumber": 1510105,
-            "journalEntryType": "UP",
-            "customerId": 26,
-            "id": 2,
-            "name": "654654",
-            "surname1": "",
-            "surname2": "",
-            "phone": 0,
-            "email": "",
-            "orderNotifications": false,
-            "orderNotificationEmail": "",
-            "orderNotificationByEmail": false,
-            "orderNotificationBySms": false,
-            "orderNotificationByPush": false,
-            "deleted": true,
-            "object": "SALQ9U",
-            "objectSchema": "IQSFCOMUN"
-        }
-    ]';
-
-    private $operatorData = [
-        'sequenceNumber' => 1510105,
-        'journalEntryType' => 'UP',
-        'customerId' => 26,
-        'id' => 2,
-        'name' => '654654',
-        'surname1' => '',
-        'surname2' => '',
-        'phone' => 0,
-        'email' => '',
-        'orderNotifications' => false,
-        'orderNotificationEmail' => '',
-        'orderNotificationByEmail' => false,
-        'orderNotificationBySms' => false,
-        'orderNotificationByPush' => false,
-        'deleted' => true,
-        'object' => 'SALQ9U',
-        'objectSchema' => 'IQSFCOMUN'
-    ];
 
     public function setUp(): void
     {
@@ -69,12 +28,14 @@ class ApiOperatorsDataSourceTest extends TestCase
     #[Test]
     public function getsOperators(): void
     {
+        $operatorTestDataBuilder = new OperatorTestDataBuilder();
         $expectedResponse = [];
-        $apiResponse = new Response(200, [], $this->operatorDataApi);
+        $expectedResponse[] = $operatorTestDataBuilder->withId(3)->build();
+        $operatorsDataApi = $operatorTestDataBuilder->withId(3)->toApiResponse();
+        $apiResponse = new Response(200, [], $operatorsDataApi);
         $this->client
             ->expects('get')
             ->andReturn($apiResponse);
-        $expectedResponse[] = new Operator($this->operatorData);
 
         $response = $this->apiDataSource->getOperators();
